Remove extra quote in bash invocation
Currently, we use the following bash invocation to run the linter:

ament_<linter> $(colcon list --packages-select 'a b c')

The use of "colcon list" is to convert the list of package names into
packages paths that the linter expects.
I.e. we can run ament_copyright /path/to/pkg/a /path/to/pkg/b, but not
ament_copyright a b.

This fails because colcon list is looking for a single package called 'a b c'.
As there is no package called 'a b c'. What we want is instead selecting
package, "a", "b", and "c". This can be done by removing the single quotes
around the package list:

ament_<linter> $(colcon list --packages-select a b c)

Unfortunately, "colcon list" returns 0 (success), even if it fails to
match anything. Therefore, there is no easy way to detect this issue
with the current implementation. An alternative solution could be to
parse stderr, which is cumbersome, and brittle. For now, we chose to
only fix the issue, without introducing a test or a safety mechanism to
prevent regressions from happening.

Another problem is that ament_<linter> "/path/to/a /path/to/b" fails due
to the same issue, those extra quotes have also been removed.

Finally, this fixes the test. Until now, the test was not executed
properly, but it was never noticed because colcon fails silently
if no packages are matched. This requires the introduction of a new
parameter workspace-director allowing the linters to rely on a colcon
workspace located into a custom directory.

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>